requirements.txt
chromadb
pypdf
openai
langchain
langchain-community
pandas
numpy
python-dotenv
tiktoken